Update to luatex 1.17.0 (2023-04-29)

+++ src/mplib/src/mpsvgout.h
@@ -22,8 +22,9 @@ int level;
 
 /*:20*//*27:*/
 
-integer dx;
-integer dy;
+double dx;
+double dy;
+
 
 /*:27*//*56:*/
